worker/requirements.txt
celery==5.3.6
redis==5.0.1
modelscope==1.13.2
diffusers==0.27.2
transformers==4.38.2
torch==2.2.0
torchvision==0.17.0
opencv-python==4.9.0.80
imageio==2.34.0
imageio-ffmpeg==0.4.9
numpy==1.26.4
Pillow==10.2.0
